rename amazon.d.ts -> alexa.d.ts
